backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv==1.0.1
httpx>=0.27.2
google-adk>=1.27.0
pydantic>=2.9.2
websockets>=13.1
